requirements.txt
stem-splitter
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.7.0+cu128
torchaudio==2.7.0+cu128
tensorboard
torch_optimizer
audiomentations
pyyaml
dlchordx
einops
tqdm
soundfile
tokenizers
pytorch-crf
